+++ tests/projects/common/settings.py
@@ -61,6 +61,7 @@ INSTALLED_SOCIALACCOUNT_APPS: tuple[str, ...] = (
     "allauth.socialaccount.providers.instagram",
     "allauth.socialaccount.providers.jupyterhub",
     "allauth.socialaccount.providers.kakao",
+    "allauth.socialaccount.providers.klaviyo_oauth2",
     "allauth.socialaccount.providers.lemonldap",
     "allauth.socialaccount.providers.lichess",
     "allauth.socialaccount.providers.line",